utl: change float to double since SWIG does a cast to double which casues convertion issues, add checks for inf and nan and record those as strings + test cases to check, but cannot enable due to missing close_metrics command

+++ src/utl/src/LoggerCommon.cpp
@@ -94,7 +94,7 @@ void metric_integer(const char* metric, const int value)
   logger->metric(metric, value);
 }
 
-void metric_float(const char* metric, const float value)
+void metric_float(const char* metric, const double value)
 {
   Logger* logger = getLogger();
   logger->metric(metric, value);
